[FEAT]: ViewController가 커짐에 있어 JoinUpPageViewController의 Extension추가
